// packages/domains/vault/vault-contracts/src/vault-items-crud/types/vault-item/vault-item-type-enum.ts
export enum VaultItemType {
  Address = "KWAddress",
  BankAccount = "KWBankStatement",
  Company = "KWCompany",
  Credential = "KWAuthentifiant",
  DriversLicense = "KWDriverLicence",
  Email = "KWEmail",
  FiscalId = "KWFiscalStatement",
  IdCard = "KWIDCard",
  Identity = "KWIdentity",
  Passkey = "KWPasskey",
  Passport = "KWPassport",
  PaymentCard = "KWPaymentMean_creditCard",
  Phone = "KWPhone",
  Secret = "KWSecret",
  SecureNote = "KWSecureNote",
  SocialSecurityId = "KWSocialSecurityStatement",
  Website = "KWPersonalWebsite",
}